[WCMSFEQ-1063] Roll ContentPage.js into Common.js (#199)
* Roll ContentPage into Common.js - Order of execution maintained

* Combine DOM ready blocks
