+++ CancerGov/release_notes/frontend-2018-07-20-july-sprint.md
@@ -1,3 +1,9 @@
 #Frontend-2018-07-20: FEQ Bucket Release (508 and Architecture Tickets)
 
 TBD...
+
+##  [WCMSFEQ-1063] Roll ContentPage into Common
+
+Instead of two global js files being requested on each page, we will only have one. Other tickets will ideally pare this file down.
+
+Content changes are required to make sure all the Percussion templates are no longer slotting in ContentPage.
